fix typos, move paddle.tensor.kron back to fluid.layers.kron, test=release/2.0
